src/requirements.txt
aws-wsgi==0.2.7
Flask==2.0.1
awslambdaric==1.2.0
Django==2.2.17
djangorestframework==3.12.4
